seems like its only working with Hibernate 4.x

+++ src/main/java/assignment8/dataServices/UserService.java
@@ -40,7 +40,7 @@ public class UserService {
 
     @POST
     public Response createUser() {
-        final Long id = UserManager.getInstance().createUser();
+        final int id = UserManager.getInstance().createUser();
 
         final URI locationURI = uriInfo.getAbsolutePathBuilder().path(Long.toString(id)).build();
